split Network and HGF classes
