+++ src/pyhgf/model/__init__.py
@@ -1,0 +1,4 @@
+from .hgf import HGF
+from .network import Network
+
+__all__ = ["HGF", "Network"]
